requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
pytest
pytest-asyncio